% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_score.R
\name{vsm_score}
\alias{vsm_score}
\title{vsm_score}
\usage{
vsm_score(df, query, accuracy = 0.5)
}
\arguments{
\item{df}{a dataframe including text documents to be used as search corpora}

\item{query}{a string of key words to be searched for}

\item{accuracy}{an accuracy threshold. Results with matching scores lower than this threshold will be discarded.}
}
\value{
This function returns a list of items that correspond to the search keys. Each one includes dataf rames of those documents from the corpora that have a matching score at or above the accuracy threshold, as well as information on the variable associated and its location inside the Microdata Library.
}
\description{
This is a helper function. Called internally by \code{WBquery()}, it implements a search engine for text documents based on a simple \href{https://en.wikipedia.org/wiki/Vector_space_model}{vector-space-model}. It is broadly based on multiple online tutorials, some of which can be found \href{https://rpubs.com/ftoresh/search-engine-Corpus}{here}, \href{https://www.r-bloggers.com/2013/03/build-a-search-engine-in-20-minutes-or-less/}{here}, and \href{https://gist.github.com/sureshgorakala/c990c3cd681b7cecdf57ef8a2ce42005}{here}.
}
\examples{
# No examples are provided for this helper function.
}
