% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBdownload.R
\name{WBdownload}
\alias{WBdownload}
\title{Title}
\usage{
WBdownload(
  item.id,
  dir = getwd(),
  unzip = TRUE,
  user.email,
  user.password,
 
    user.abstract = "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?"
)
}
\arguments{
\item{item.id}{unique numeric identifier for a WB Microdata Library item.}

\item{dir}{a directory on the local machine to write files to (default is \code{getwd()}).}

\item{unzip}{a logical indicating whether downloads should be unzipped (default is \code{TRUE}).}

\item{user.email}{a valid email address associated with a registered user account at the WB Microdata Library.}

\item{user.password}{a valid password associated with the same registered user account at the WB Microdata Library as \code{user.email}.}

\item{user.abstract}{a short description of the purpose for which the data is being downloaded (defaults to Sect. 1.10.32 of M.T. Cicero's "de Finibus Bonorum et Malorum" as a placeholder).}
}
\value{
This function does not return any element.
}
\description{
Title
}
\examples{
\dontrun{
WBdownload(item.id = 1234,
user.email = "user@example.com",
user.password = "a.VerY_DiffiCULT-PasSWoRD",
user.abstract = "Writing a very innovative paper on something rather important."
dir = "C:/Users/myusername/Desktop/",
unzip = FALSE)
}

}
