% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBquery.R
\name{WBquery}
\alias{WBquery}
\title{WBquery}
\usage{
WBquery(dest, key, iso, col)
}
\arguments{
\item{dest}{File destination (default is current working directory)}

\item{key}{A vector of keywords used in the quary}

\item{iso}{Iso Alpha 3 country codes in capital letters (ABW, AFG, AGO,...)}

\item{col}{ID of one or more World Bank Microdata Library collection(s). There are 23 choices: \code{c("afrobarometer", "datafirst","dime", "microdata_rg", "enterprise_surveys", "FCV", "global-findex", "ghdx", "hfps", "impact_evaluation", "ipums", "lsms", "dhs", "mrs", "MCC", "pets", "sdi", "step", "sief", "COS", "MICS", "unhcr", "WHO")}.}

\item{out}{One of \code{c("search", "get")}}
}
\value{
If \code{out = "search"}, \code{WBquery()} returns a list of items that match the search criteria. If \code{out = "get"}, \code{WBquery()} downloads all items that match the serach criteria to the current working directory, or the directory specified by \code{dest}.
}
\description{
WBquery
}
