% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBget.R
\name{WBget}
\alias{WBget}
\title{WBget}
\usage{
WBget(id)
}
\arguments{
\item{id}{vector of dataset identifiers}
}
\value{
the function creates a new folder and saves the requested datasets there.
}
\description{
WBget
}
\examples{
ids <- c("AFG_2020_VOLREP_v01_M", "AFG_2021_WBCS_v01_M", "AFG_2019_LIS_v01_M",
"AFG_2019_VOLREP_v01_M", "AFG_2018_SPA_v01_M", "AFG_2018_VOLREP_v01_M",
"AFG_2018_WBCS_v01_M","AFG_2017_FINDEX_v02_M", "AFG_2017_SABER-SD_v01_M",
"AFG_2017_SEA-HER_v01_M", "AFG_2017_SEA-KHO_v01_M", "AFG_2016_VOLREP_v01_M")

WBget(ids)
}
